# internal/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Elastic Maintainer API",
    "version": "v1alpha1",
    "description": "Initial web-first API contract. Endpoints marked with 501 responses are registered skeletons whose domain execution lands in later phases."
  },
  "security": [
    {
      "sessionCookie": []
    },
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/sources": {
      "get": {
        "operationId": "listSources",
        "responses": {
          "200": {
            "description": "Mounted source inventory",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/APIError"
          },
          "401": {
            "$ref": "#/components/responses/APIError"
          },
          "403": {
            "$ref": "#/components/responses/APIError"
          },
          "503": {
            "$ref": "#/components/responses/APIError"
          },
          "422": {
            "$ref": "#/components/responses/APIError"
          },
          "409": {
            "$ref": "#/components/responses/APIError"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/PageToken"
          }
        ]
      }
    },
...
    "/api/v1/targets": {
      "get": {
        "operationId": "listTargets",
        "responses": {
          "200": {
            "description": "Target inventory",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/APIError"
          },
          "401": {
            "$ref": "#/components/responses/APIError"
          },
          "403": {
            "$ref": "#/components/responses/APIError"
          },
          "503": {
            "$ref": "#/components/responses/APIError"
          },
          "422": {
            "$ref": "#/components/responses/APIError"
          },
          "409": {
            "$ref": "#/components/responses/APIError"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/PageToken"
          }
        ]
      }
    },
...
    "/api/v1/validations": {
      "get": {
        "operationId": "listValidationJobs",
        "responses": {
          "200": {
            "description": "Validation jobs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/APIError"
          },
          "401": {
            "$ref": "#/components/responses/APIError"
          },
          "403": {
            "$ref": "#/components/responses/APIError"
          },
          "503": {
            "$ref": "#/components/responses/APIError"
          },
          "409": {
            "$ref": "#/components/responses/APIError"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/PageToken"
          }
        ]
      },
      "post": {
        "operationId": "createValidationJob",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidationCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "$ref": "#/components/responses/JobAccepted"
          },
          "400": {
            "$ref": "#/components/responses/APIError"
          },
          "401": {
            "$ref": "#/components/responses/APIError"
          },
          "403": {
            "$ref": "#/components/responses/APIError"
          },
          "409": {
            "$ref": "#/components/responses/APIError"
          },
          "415": {
            "$ref": "#/components/responses/APIError"
          },
          "429": {
            "$ref": "#/components/responses/APIError"
          },
          "503": {
            "$ref": "#/components/responses/APIError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-elastic-maintainer-session"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "OIDC JWT access token, bounded to 16 KiB and validated with RS256 or ES256 against the configured pinned issuer, client audience, JWKS transport, and claim-to-role mapping. Bearer tokens do not require a browser nonce. Duplicate Authorization headers or any simultaneous application-session cookie are rejected."
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 8,
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{7,127}$"
        }
      },
...
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      },
      "PageToken": {
        "name": "pageToken",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 512
        }
      },
...
    },
    "responses": {
      "APIError": {
        "description": "API error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
      "JobAccepted": {
        "description": "Job accepted",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/JobAcceptedResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "APIVersion": {
        "type": "string",
        "enum": [
          "elastic-maintainer/v1alpha1"
        ]
      },
      "SafeID": {
        "type": "string",
        "minLength": 1,
        "maxLength": 64,
        "pattern": "^[A-Za-z0-9_-]+$"
      },
...
      "ValidationCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "resourceSetIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SafeID"
            },
            "uniqueItems": true,
            "maxItems": 100
          },
          "targetIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SafeID"
            },
            "uniqueItems": true,
            "maxItems": 100
          }
        }
      },
...
      "Job": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "type",
          "status",
          "createdAt",
          "actorSubject",
          "requestId"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/SafeID"
          },
          "type": {
            "type": "string",
            "enum": [
              "validation",
              "plan",
              "apply"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "canceled",
              "interrupted"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "actorSubject": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "failureCode": {
            "type": "string"
          }
        }
      },
...
      "SourceListResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "apiVersion",
          "sources"
        ],
        "properties": {
          "apiVersion": {
            "$ref": "#/components/schemas/APIVersion"
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceSummary"
            }
          },
          "nextPageToken": {
            "type": "string"
          }
        }
      },
...
      "TargetListResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "apiVersion",
          "targets"
        ],
        "properties": {
          "apiVersion": {
            "$ref": "#/components/schemas/APIVersion"
          },
          "targets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TargetSummary"
            }
          },
          "nextPageToken": {
            "type": "string"
          }
        }
      },
...
      "JobAcceptedResponse": {
        "type": "object",
        "required": [
          "apiVersion",
          "job"
        ],
        "properties": {
          "apiVersion": {
            "$ref": "#/components/schemas/APIVersion"
          },
          "job": {
            "$ref": "#/components/schemas/Job"
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "apiVersion",
          "error"
        ],
        "properties": {
          "apiVersion": {
            "$ref": "#/components/schemas/APIVersion"
          },
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              }
            }
          }
        }
      },
      "Digest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "algorithm",
          "version",
          "value"
        ],
        "properties": {
          "algorithm": {
            "type": "string",
            "enum": [
              "sha256"
            ]
          },
          "version": {
            "type": "string",
            "enum": [
              "v1"
            ]
          },
          "value": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          }
        }
      },
      "RevisionProvenance": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "relativePath",
          "value"
        ],
        "properties": {
          "relativePath": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "SourceSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "desiredDigest",
          "fileCount",
          "resourceCount"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/SafeID"
          },
          "revision": {
            "$ref": "#/components/schemas/RevisionProvenance"
          },
          "desiredDigest": {
            "$ref": "#/components/schemas/Digest"
          },
          "fileCount": {
            "type": "integer",
            "minimum": 0
          },
          "resourceCount": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "TargetSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "identity",
          "labels",
          "resourceSetID",
          "desiredDigest",
          "resourceCount"
        ],
        "properties": {
          "identity": {
            "type": "object"
          },
          "labels": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "resourceSetID": {
            "$ref": "#/components/schemas/SafeID"
          },
          "revision": {
            "type": "string"
          },
          "desiredDigest": {
            "$ref": "#/components/schemas/Digest"
          },
          "resourceCount": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "ValidationListResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "apiVersion",
          "jobs"
        ],
        "properties": {
          "apiVersion": {
            "$ref": "#/components/schemas/APIVersion"
          },
          "jobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Job"
            }
          },
          "nextPageToken": {
            "type": "string"
          }
        }
      },
...
    }
  }
}